move waveform to jsui from javascript
